requirements.txt
google-generativeai
PyMuPDF
pandas
python-dotenv
tk  # Para el selector de carpetas (generalmente viene incluido con Python)